requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
scipy>=1.10.0
pyyaml>=6.0
joblib>=1.2.0
lightgbm>=3.3.0
streamlit>=1.20.0
plotly>=5.13.0
